tests/wpt/tests/shadow-dom/declarative/declarative-parser-interaction.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Declarative Shadow DOM</title>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel="help" href="https://github.com/whatwg/dom/issues/831">
<link rel="help" href="https://crbug.com/1203645">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helpers.js"></script>


This test should not crash, and there should be two lines of text visible below.
<x-1>
    <template shadowrootmode="open">
        <style>
            @import 'non-existent.css';
        </style>
        <slot></slot>
    </template>
    <p>Line 1</p>
</x-1>

<x-2>
    <template shadowrootmode="open">
        <slot></slot>
    </template>
    <p>Line 2</p>
</x-2>

<script>
  window.onload = function() {
    const x1 = document.querySelector('x-1');
    const x2 = document.querySelector('x-2');
    test(() => {
      assert_true(!!x1);
      assert_true(!!x2);
    }, 'Declarative Shadow DOM: Test for crashes and improper parsing');
  }
</script>
